% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{anpan}
\alias{anpan}
\title{Run anpan}
\usage{
anpan(
  bug_file,
  meta_file,
  out_dir,
  model_type = "horseshoe",
  covariates = c("age", "gender"),
  outcome = "crc",
  omit_na = FALSE,
  filtering_method = "kmeans",
  discard_absent_samples = TRUE,
  annotation_file = NULL,
  plot_ext = "png",
  save_filter_stats = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bug_file}{path to a gene family file (usually probably from HUMAnN)}

\item{meta_file}{path to a metadata tsv}

\item{out_dir}{path to the desired output directory}

\item{model_type}{either "horseshoe", "glm", or "fastglm"}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{omit_na}{logical indicating whether to omit incomplete cases}

\item{discard_absent_samples}{logical indicating whether to discard samples when a bug is labelled as completely absent}

\item{save_filter_stats}{logical indicating whether to save filter statistics}

\item{...}{arguments to pass to brms::brm()}
}
\description{
Run the anpan model on a single bug
}
\details{
The specified metadata file must contain columns matching "sample_id"
  and the specified covariates and outcome variables.
}
