% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_outcome.R
\name{anpan_latent_outcome}
\alias{anpan_latent_outcome}
\title{Estimate the latent outcome model}
\usage{
anpan_latent_outcome(
  model_mat,
  dist_mat,
  iter_sampling = 1000,
  iter_warmup = 1000,
  sigma_factor = 5,
  verbose = TRUE,
  algorithm = "mcmc",
  ...
)
}
\arguments{
\item{model_mat}{a model matrix. Should not have an intercept column.}

\item{dist_mat}{a distance matrix.}

\item{iter_sampling}{the number of post-warmup draws to take}

\item{iter_warmup}{the number of warmup draws}

\item{sigma_factor}{higher --> lower prior spread on lm residuals}

\item{verbose}{verbose setting}

\item{algorithm}{either "mcmc" or "variational"}

\item{...}{other parameters to pass to cmdstanr::sample()}
}
\value{
a named list. draw_df is set of MCMC draws of each parameter.
  summary_df is the posterior summary. sample_order is a tibble containing
  the sample order used by the model. y_scale is the empirically estimated
  prior scale of the latent y values and residiuals. term calls is a data
  frame of ANOVA-like "calls" for each term in the model formula, indicating
  whether any of its associated coefficients are non-zero by 90% interval
  excluding 0.
}
\description{
Run a latent outcome model on a distance matrix and a model
  matrix.
}
\details{
The variational fit is faster but biased.
}
