% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.R
\name{anpan_pglmm}
\alias{anpan_pglmm}
\title{Run a phylogenetic generalized linear mixed model}
\usage{
anpan_pglmm(
  meta_file,
  tree_file,
  trim_pattern = NULL,
  covariates = NULL,
  plot_cov_mat = TRUE,
  outcome = "age",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{meta_file}{path to a metadata tsv}

\item{tree_file}{path to a .tre file}

\item{trim_pattern}{optional pattern to trim from tip labels of the tree}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{...}{other arguments to pass to brms::brm}
}
\description{
Run a phylogenetic generalized linear mixed model
}
\details{
the tip labels of the tree must be the sample ids from the metadata.
  You can use the \code{trim_pattern} argument to automatically trim off any
  consistent pattern from the tip labels if necessary. The dots can be used
  to pass cores=4 to brms to make the chains run in parallel.

  The prior for the intercept is a normal distribution centered on the mean
  of the outcome variable with a standard deviation of 3*sd(outcome variable)
}
