% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwy_ranef.R
\name{anpan_pwy_ranef}
\alias{anpan_pwy_ranef}
\title{Estimate a species-pathway abundance random effects model}
\usage{
anpan_pwy_ranef(bug_pwy_dat, group_ind = "crc", ...)
}
\arguments{
\item{bug_pwy_dat}{a data frame with a row for each observation and columns "pwy",
"log10_species_abd", "log10_pwy_abd", and a group indicator column named according to the
\code{group_ind} argument}

\item{group_ind}{a character giving the name of the column for the 0/1 group indicator variable
in \code{bug_pwy_dat}}

\item{...}{other arguments to pass to cmdstanr::sample()}
}
\value{
a list containing the CmdStanMCMC object of the model fit and the pwy-index map
}
\description{
Fit a model of the form pwy_abd ~ species_abd + (1|pwy) + (0+group|pwy) for a single
  bug
}
\details{
The priors are as follows: A) student_t(3, mean(pwy_abd), 2.5) on the intercept at the
  mean species abundance. B) half student_t(3, 0, 2.5) on the residual noise C) 0-centered normal
  priors on pathway specific effects C1) half student_t(5,0,2.5) on the SD of pathway intercepts
  C2) half standard normal on the SD of group-specific effects.

  The pathway index is generated by converting the pwy column to a factor and then to the
  corresponding integer index.

  The group_ind column should be numeric with values in {0,1}

  The main parameter of interest is the \code{pwy_effects} vector. When
}
\seealso{
[cmdstanr::CmdStanMCMC()]
}
