% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwy_ranef.R
\name{anpan_pwy_ranef_batch}
\alias{anpan_pwy_ranef_batch}
\title{Fit the pathway random effects model for multiple bugs}
\usage{
anpan_pwy_ranef_batch(bug_pwy_dat, group_ind = "crc", ...)
}
\arguments{
\item{bug_pwy_dat}{a data frame with a row for each observation and columns "pwy",
"log10_species_abd", "log10_pwy_abd", and a group indicator column named according to the
\code{group_ind} argument}

\item{group_ind}{a character giving the name of the column for the 0/1 group indicator variable
in \code{bug_pwy_dat}}

\item{...}{other arguments to pass to cmdstanr::sample()}
}
\value{
a tibble of row-binded anpan_pwy_ranef results
}
\description{
Fit the pathway random effects model for multiple bugs
}
\details{
In addition to the column requirements of \code{anpan_pwy_ranef()}, the input data frame
  \code{bug_pwy_dat} here must also contain a variable called "bug" which gives a unique
  identifier for each bug.
}
