% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{anpan_repeated_measures}
\alias{anpan_repeated_measures}
\title{Use repeated measures to refine the gene model}
\usage{
anpan_repeated_measures(
  subject_sample_map,
  bug_dir,
  meta_file,
  out_dir,
  model_type = "fastglm",
  covariates = c("age", "gender"),
  outcome = "crc",
  omit_na = FALSE,
  filtering_method = "kmeans",
  discard_poorly_covered_samples = TRUE,
  skip_large = TRUE,
  save_fit = TRUE,
  annotation_file = NULL,
  save_filter_stats = TRUE,
  verbose = TRUE,
  plot_result = TRUE,
  plot_ext = "png",
  q_threshold = NULL,
  n_top = 50,
  width = 10,
  height = 8,
  ...
)
}
\arguments{
\item{subject_sample_map}{a data frame between the mapping between subject_id and sample_id}

\item{bug_dir}{a directory of gene family files}

\item{meta_file}{path to a metadata tsv}

\item{out_dir}{path to the desired output directory}

\item{model_type}{either "horseshoe" or "fastglm"}

\item{covariates}{character vector of covariates to include in the model}

\item{outcome}{character string of the outcome variable}

\item{omit_na}{logical indicating whether to omit incomplete cases of the metadata}

\item{filtering_method}{method to use for filtering samples. Either "kmeans" or "none"}

\item{discard_poorly_covered_samples}{logical indicating whether to discard samples where the genes of a bug are poorly covered}

\item{skip_large}{logical indicating whether to skip bugs with over 5k genes. Only used when
model_type = "horseshoe".}

\item{save_fit}{logical indicating whether to save horseshoe fit objects. Only used when
model_type = "horseshoe".}

\item{annotation_file}{a path to a file giving annotations for each gene}

\item{save_filter_stats}{logical indicating whether to save filter statistics}

\item{plot_result}{logical indicating whether or not to plot the results}

\item{plot_ext}{extension to use for plots}

\item{q_threshold}{FDR threshold to use for inclusion in the plot.}

\item{n_top}{number of top elements to show from the results}

\item{width}{width of saved plot in inches}

\item{height}{height of saved plot in inches}

\item{...}{arguments to pass to [cmdstanr::sample()] if applicable}
}
\description{
Use repeated measures to refine the gene model
}
\details{
This function performs the standard anpan filtering on all samples, then uses the
  subject-sample map to compute the proportion of samples with the bug. This gives a gene
  _proportion_ matrix (instead of a presence/absence matrix) which is then passed to
  \code{anpan_batch(filtering_method = "none", discretize_inputs = FALSE)}.

  In cases where subject metadata varies by sample, the mean is taken if the variable is numeric,
  otherwise it is tabulated and the most frequent category is selected as the subject-level
  metadata value. This tabulation will respect factor ordering if you'd like to alter the value
  selected in the event of ties.
}
