% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.R
\name{anpan_subjectwise_pglmm}
\alias{anpan_subjectwise_pglmm}
\title{Fit a subject-wise PGLMM}
\usage{
anpan_subjectwise_pglmm(
  tree_file,
  meta_file,
  subject_sample_map,
  outcome,
  covariates = NULL,
  out_dir = NULL,
  trim_pattern = NULL,
  omit_na = FALSE,
  family = "gaussian",
  show_plot_cor_mat = TRUE,
  show_plot_tree = TRUE,
  show_post = TRUE,
  show_yrep = FALSE,
  save_object = FALSE,
  verbose = TRUE,
  loo_comparison = TRUE,
  reg_noise = TRUE,
  reg_gamma_params = c(1, 2),
  plot_ext = "pdf",
  beta_sd = NULL,
  sigma_phylo_scale = 0.333,
  ...
)
}
\arguments{
\item{tree_file}{either a path to a tree file readable by \code{\link[ape:read.tree]{ape::read.tree()}} or an object of
class "phylo" that is already read into R. Ignored if \code{cor_mat} is supplied.}

\item{meta_file}{either a data frame or a file. Can provide covariate and outcome variables for
either subjects or samples.}

\item{subject_sample_map}{a data frame giving a mapping between sample_id (which must match the
leaves of the tree up to the trim_pattern) and subject_id}

\item{outcome}{the name of the outcome variable}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{out_dir}{if saving, directory where to save}

\item{trim_pattern}{optional pattern to trim from tip labels of the tree}

\item{omit_na}{logical indicating whether to omit incomplete cases of the metadata}

\item{family}{string giving the name of the distribution of the outcome variable (usually
"gaussian" or "binomial")}

\item{show_plot_cor_mat}{show a plot of the correlation matrix derived from the tree}

\item{show_plot_tree}{show a plot of the tree overlaid with the outcome.}

\item{show_post}{show a plot of the tree overlaid with the outcome and posterior distribution on
phylogenetic effects.}

\item{show_yrep}{show a plot of the tree overlaid with the outcome and the posterior predictive
distribution for each observation if plotting the tree}

\item{save_object}{logical indicating whether to save the model fit object}

\item{loo_comparison}{logical indicating whether to compare the phylogenetic model against a base
model (without the phylogenetic term) using \code{\link[loo:loo_compare]{loo::loo_compare()}}}

\item{reg_noise}{logical indicating whether to regularize the ratio of sigma_phylo to sigma_resid
with a Gamma prior}

\item{reg_gamma_params}{the shape and rate parameters of the Gamma prior on the noise term ratio.
Default: c(1,2)}

\item{plot_ext}{extension to use when saving plots}

\item{beta_sd}{prior standard deviation parameters on the normal distribution for each covariate
in the GLM component}

\item{sigma_phylo_scale}{standard deviation of half-normal prior on \code{sigma_phylo} for
logistic PGLMMs when \code{family = 'binomial'}. Increasing this value can easily lead to
overfitting.}

\item{...}{other arguments to pass to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}}
}
\description{
If you have a dataset where multiple samples come from the same individual, running
  a PGLMM on all the samples without accounting for this will return a strong spurious signal
  because samples from the same individual will (usually) be right next to each other on the tree
  and all show the same outcome value. This function uses a subject-sample map input to aggregate
  samples by subject, derive a subject-wise correlation matrix, then run a PGLMM on that.
}
\details{
The \code{meta_file} must contain at least one column named "sample_id" or "subject_id".
  If the metadata is inferred to be provided by sample, representative covariate and outcome
  variable values are selected for each subject in the manner described in
  \code{?anpan_repeated_measures()}
}
\seealso{
\code{\link[=anpan_pglmm]{anpan_pglmm()}} \code{\link[=anpan_repeated_measures]{anpan_repeated_measures()}}
}
