% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{anpan_vignette}
\alias{anpan_vignette}
\title{Open the anpan vignette}
\usage{
anpan_vignette()
}
\description{
This function checks if you have the vigette built, then opens
it with utils::RShowDoc() if you do. This function exists just to make it a
little easier for users to find their way to the vignette.
}
