% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade_effects.R
\name{compute_clade_effects}
\alias{compute_clade_effects}
\title{Compute a clade effect}
\usage{
compute_clade_effects(
  clade_members,
  anpan_pglmm_result,
  plot_difference = TRUE
)
}
\arguments{
\item{clade_members}{a character vector listing members of the clade of interest}

\item{anpan_pglmm_result}{a result list from \code{anpan_pglmm()}}

\item{plot_difference}{logical indicating whether to plot 50% and 98% posterior intervals for the
three computed variables}
}
\value{
a list of two tibbles
}
\description{
This function computes the average phylogenetic effect of members of a user-defined
  clade against that of all non-members.
}
