% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{draw_mvnorm}
\alias{draw_mvnorm}
\title{Draw a sample from a multivariate normal distribution}
\usage{
draw_mvnorm(mean = NULL, Sigma = NULL, L = NULL)
}
\arguments{
\item{mean}{mean vector, defaults to 0 vector}

\item{Sigma}{covariance matrix}

\item{L}{optional cholesky factor of \code{Sigma}}
}
\value{
A vector giving one draw from the specified multivariate normal
  distribution.
}
\description{
Draw a sample from a multivariate normal distribution with
  specified mean and covariance.
}
\details{
One of \code{Sigma} or \code{L} must be provided. Pre-computing L
  with \code{L = t(chol(Sigma))} will make calling this function repeatedly
  much faster.
}
