% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_batch}
\alias{filter_batch}
\title{Filter a batch of files}
\usage{
filter_batch(
  bug_dir,
  meta_file,
  filter_stats_dir,
  pivot_wide = TRUE,
  minmax_thresh = 5,
  covariates = NULL,
  outcome = NULL,
  filtering_method = "kmeans",
  discretize_inputs = TRUE,
  discard_poorly_covered_samples = TRUE,
  omit_na = FALSE,
  plot_ext = "pdf",
  verbose = TRUE
)
}
\arguments{
\item{bug_dir}{a directory of gene family files}

\item{meta_file}{path to a metadata tsv}

\item{filter_stats_dir}{directory to save filtering statistics to}

\item{pivot_wide}{logical indicating whether to return data in wide format}

\item{minmax_thresh}{genes must have at least this many (or N - this many)
non-zero observations or else be discarded}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{filtering_method}{either "kmeans" or "none"}

\item{discretize_inputs}{logical indicating whether to discretize the input abundance
measurements (0/nonzero --> FALSE/TRUE) before passing them to the modelling function}

\item{discard_poorly_covered_samples}{logical indicating whether to discard samples where the genes of a bug are poorly covered}

\item{omit_na}{logical indicating whether to omit incomplete cases of the metadata}

\item{plot_ext}{extension to use for plots}
}
\value{
A list of filtered data frames
}
\description{
This function applies anpan::read_and_filter() to a set of files.
}
