% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_gf}
\alias{filter_gf}
\title{Filter a gene family file}
\usage{
filter_gf(
  gf,
  samp_stats,
  filtering_method = "kmeans",
  covariates = NULL,
  outcome = NULL,
  genomes_file = NULL,
  discard_poorly_covered_samples = TRUE,
  save_filter_stats = FALSE,
  filter_stats_dir = NULL,
  plot_ext = "pdf",
  bug_name = NULL
)
}
\arguments{
\item{gf}{a gene family data.table}

\item{samp_stats}{a data.table of sample statistics}

\item{filtering_method}{method to use for filtering samples. Either "kmeans" or "none"}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{genomes_file}{optional file giving gene presence/absence of representative isolate genomes}

\item{discard_poorly_covered_samples}{logical indicating whether to discard samples where the genes of a bug are poorly covered}

\item{save_filter_stats}{logical indicating whether to save filter statistics}

\item{filter_stats_dir}{directory to save filtering statistics to}

\item{plot_ext}{extension to use for plots}
}
\description{
This is the function that anpan uses to filter an input
  genefamily data.table. This can be useful if you want to play around with
  different filtering options without repeatedly re-reading or checking the
  file.
}
