% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.R
\name{get_cor_mat}
\alias{get_cor_mat}
\title{Get the correlation matrix of a tree}
\usage{
get_cor_mat(bug_tree)
}
\arguments{
\item{bug_tree}{a tree object of class "phylo"}
}
\description{
Get the correlation matrix of a tree
}
\seealso{
[ape::read.tree()]
}
