% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_qc.R
\name{get_genome_intersect_counts}
\alias{get_genome_intersect_counts}
\title{Get genome intersection counts}
\usage{
get_genome_intersect_counts(unistrain_file)
}
\description{
Get genome intersection counts
}
