% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_data_plot}
\alias{make_data_plot}
\title{Plot the data for top results}
\usage{
make_data_plot(
  res,
  covariates,
  outcome,
  model_input,
  plot_dir,
  bug_name,
  annotation_file = NULL,
  plot_ext = ".pdf",
  n_top = 50,
  q_threshold = NULL
)
}
\arguments{
\item{res}{a data frame of model results (from \code{anpan} or
\code{anpan_batch})}

\item{covariates}{character string of the covariates to show}

\item{outcome}{character string of the outcome variable}

\item{model_input}{data frame of the model input}

\item{plot_dir}{directory to write output to}

\item{bug_name}{character string giving the name to use in the title/output
file}

\item{annotation_file}{optional path file giving annotations}

\item{plot_ext}{extension to use for plots}

\item{n_top}{number of top elements to show from the results}

\item{q_threshold}{FDR threshold to use for inclusion in the plot.}
}
\description{
This funciton makes a tile plot of the top results of a fit
  alongside another tile plot showing the covariates included. Optional
  annotations can be included.
}
\details{
If included, \code{annotation_file} must be a tsv with two columns:
  "gene" and "annotation".

  \code{n_top} is ignored if \code{q_threshold} is specified.
}
