% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{make_line_plot}
\alias{make_line_plot}
\title{Make a filter-labelled line plot}
\usage{
make_line_plot(
  bug_file = NULL,
  meta_file = NULL,
  covariates,
  outcome,
  fgf = NULL,
  bug_name = NULL,
  plot_dir = NULL
)
}
\arguments{
\item{bug_file}{a path to a gene family file}

\item{meta_file}{a path to the corresponding metadata file}

\item{covariates}{covariates to account for}

\item{outcome}{the name of the outcome variable}

\item{fgf}{a filtered gene family data frame}

\item{bug_name}{name of the bug}
}
\description{
Make a filter-labelled line plot
}
\details{
The required input is either \itemize{ \item{the gene family file
  and the metadata file} \item{OR a pre-filtered gene family file}}
}
