% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.R
\name{olap_tree_and_meta}
\alias{olap_tree_and_meta}
\title{Overlap a tree and metadata}
\usage{
olap_tree_and_meta(
  tree_file,
  meta_file,
  covariates,
  outcome,
  omit_na = FALSE,
  trim_pattern,
  verbose = TRUE
)
}
\arguments{
\item{tree_file}{either a path to a tree file readable by \code{\link[ape:read.tree]{ape::read.tree()}}
or an object of class "phylo" that is already read into R}

\item{meta_file}{either a data frame of metadata or a path to file containing
the metadata}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{omit_na}{logical indicating whether to omit incomplete cases}

\item{trim_pattern}{optional pattern to trim from tip labels of the tree}
}
\description{
Overlap a tree's leaves with the observations in metadata,
  returning an intersected tree and metadata data.table.
}
