% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cor_mat}
\alias{plot_cor_mat}
\title{Plot a correlation matrix
Generate a colorful heatmap visualization of a correlation matrix}
\usage{
plot_cor_mat(cor_mat, bug_name = NULL)
}
\arguments{
\item{cor_mat}{a correlation matrix}

\item{bug_name}{a string giving the name of the bug}
}
\value{
a ggplot of the correlation matrix
}
\description{
Plot a correlation matrix
Generate a colorful heatmap visualization of a correlation matrix
}
