% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_qc.R
\name{plot_genome_intersection_histograms}
\alias{plot_genome_intersection_histograms}
\title{Intersection histograms}
\usage{
plot_genome_intersection_histograms(
  unistrain_file,
  out_dir = NULL,
  file_ext = "pdf",
  width = 10,
  height = 5
)
}
\description{
Intersection histograms
}
