% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_qc.R
\name{plot_genome_intersections}
\alias{plot_genome_intersections}
\title{Plot genome intersections}
\usage{
plot_genome_intersections(
  unistrain_file,
  out_dir = NULL,
  file_ext = "pdf",
  width = 10,
  height = 8
)
}
\description{
Plot genome intersections
}
