% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_half_cor_mat}
\alias{plot_half_cor_mat}
\title{Plot a rotated half correlation matrix}
\usage{
plot_half_cor_mat(cor_mat, border_width = 0)
}
\arguments{
\item{cor_mat}{a correlation matrix (must have dimnames)}

\item{border_width}{width parameter of the border around each cell.}
}
\value{
a ggplot of the lower triangle of the matrix.
}
\description{
Plot the lower triangle of a correlation matrix
}
\details{
If you see a thin, pixel-width grey border around each cell, try setting border_width =
  0.1 or so, depending on your output resolution.

  No checks are made on the order of the columns. If you want the order to line up with another
  plot, you'll need to check the input manually beforehand.
}
