% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_outcome_tree}
\alias{plot_outcome_tree}
\title{Plot a tree file showing the outcome variable}
\usage{
plot_outcome_tree(
  tree_file,
  meta_file,
  covariates = c("age", "gender"),
  outcome = "crc",
  omit_na = FALSE,
  ladderize = TRUE,
  color_bars = FALSE,
  verbose = TRUE,
  trim_pattern = NULL,
  return_tree_df = FALSE
)
}
\arguments{
\item{tree_file}{either a path to a tree file readable by ape::read.tree() or an object of class "phylo" that is already read into R.}

\item{meta_file}{either a data frame of metadata or a path to file containing the metadata}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{omit_na}{logical indicating whether to omit incomplete cases of the metadata}

\item{ladderize}{logical indicating whether to run \code{\link[ape:ladderize]{ape::ladderize()}} on the tree before running
the model}

\item{color_bars}{if true, show color bars below the plot showing the covariates and outcome
variables.}

\item{trim_pattern}{optional pattern to trim from tip labels of the tree}

\item{return_tree_df}{if true, return a list containing 1) the plot, 2) the segment data frame,
and 3) the labelled terminal segment data frame. Otherwise, just return the plot.}
}
\description{
Plot a tree file, and show the outcome variable as a colored dot on the end of each
  tip.
}
\details{
Showing the covariates as color bar annotations isn't supported yet.
}
