% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwy_ranef.R
\name{plot_pwy_ranef}
\alias{plot_pwy_ranef}
\title{Plot a pathway random effects result}
\usage{
plot_pwy_ranef(
  bug_pwy_dat,
  pwy_ranef_res,
  group_ind = "crc",
  group_labels = c("ctrl", "case"),
  bug_name = NULL,
  max_pwy = 20,
  post_draws = 40
)
}
\arguments{
\item{bug_pwy_dat}{a data frame with a row for each observation and columns "pwy",
"log10_species_abd", "log10_pwy_abd", and a group indicator column named according to the
\code{group_ind} argument}

\item{group_ind}{a character giving the name of the column for the 0/1 group indicator variable
in \code{bug_pwy_dat}}

\item{max_pwy}{the maximum number of bug:pwy facets to include}
}
\description{
Plot a pathway random effects result
}
\details{
If \code{bug_name} is specified, \code{bug_pwy_dat} is first filtered to just data from
  that bug.
}
