% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwy_ranef.R
\name{plot_pwy_ranef_intervals}
\alias{plot_pwy_ranef_intervals}
\title{Plot a pathway random effects result}
\usage{
plot_pwy_ranef_intervals(pwy_ranef_res, group_ind = "crc", ncol = 1)
}
\arguments{
\item{pwy_ranef_res}{a result from \code{\link[=anpan_pwy_ranef]{anpan_pwy_ranef()}} or
\code{\link[=anpan_pwy_ranef_batch]{anpan_pwy_ranef_batch()}}}

\item{group_ind}{a character giving the name of the column for the 0/1 group indicator variable
in \code{bug_pwy_dat}}

\item{ncol}{The number of columns to use if faceting multiple bugs.}
}
\description{
Plot a pathway random effects result
}
