% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_results}
\alias{plot_results}
\title{Plot the data for top results}
\usage{
plot_results(
  res,
  covariates,
  outcome,
  model_input,
  bug_name = NULL,
  discretize_inputs = TRUE,
  plot_dir = NULL,
  annotation_file = NULL,
  cluster = "none",
  show_trees = FALSE,
  n_top = 50,
  q_threshold = 0.1,
  beta_threshold = 1,
  show_intervals = TRUE,
  max_anno_width = 70,
  width = NULL,
  height = NULL,
  plot_ext = "pdf"
)
}
\arguments{
\item{res}{a data frame of model results (from \code{anpan}) for the genes of a single bug (i.e.
the output written to *gene_terms.tsv.gz)}

\item{covariates}{character string of the covariates to show}

\item{outcome}{character string of the outcome variable}

\item{model_input}{data frame of the model input}

\item{bug_name}{character string giving the name to use in the title/output file}

\item{plot_dir}{directory to write output to}

\item{annotation_file}{optional path file giving annotations}

\item{cluster}{axis to cluster. either "none", "samples", "genes", or "both"}

\item{show_trees}{logical to show the trees for the samples (if clustered)}

\item{n_top}{number of top elements to show from the results}

\item{q_threshold}{FDR threshold to use for inclusion in the plot.}

\item{beta_threshold}{Regression coefficient threshold to use for inclusion in the plot. Set to 0
to include everything.}

\item{width}{width of saved plot in inches}

\item{height}{height of saved plot in inches}

\item{plot_ext}{extension to use for plots}
}
\description{
This funciton makes a tile plot of the top results of a fit alongside another tile
  plot showing the covariates included. Optional annotations can be included.
}
\details{
If included, \code{annotation_file} must be a tsv with two columns: "gene" and
  "annotation".

  \code{n_top} is ignored if \code{q_threshold} is specified.

  When \code{cluster = "none"}, the samples are ordered by metadata and the genes are ordered by
  statistical significance.

  When significance stars are shown, they encode the following (fairly standard) significance
  thresholds: p.value < .001 ~ ***, p.value < .01  ~ **, p.value < .05  ~ *, p.value < .1   ~ .,
  p.value < 1    ~ " "

  If applicable, the Q-value used to color the dot on the interval panel is q_global if present
  in the input and q_bug_wise otherwise. That means that you'll get different results if you
  compare the output of \code{anpan_batch()} and a manual call to \code{plot_results()} using the
  bug-wise results from the \code{model_stats/} output directory. If you'd like to replicate the
  \code{anpan_batch()} plots exactly, read in the \code{all_bug_gene_terms.tsv.gz} result from
  the top level output directory, then filter it to the bug of interest.

  Note that the beta_threshold uses the value of the estimate column directly, so it is
  interpreted according to the units of your outcome variable with a continuous outcome, and on
  the log-odds scale with a binary outcome. So the default value of 1 is pretty big for a binary
  outcome, but if the spread of your continuous outcome variable is ~5 the default value of 1
  won't exclude very much.
}
