% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a tree file showing the outcome variable}
\usage{
plot_tree(
  tree_file,
  meta_file,
  covariates = c("age", "gender"),
  outcome = "crc",
  out_dir = NULL,
  omit_na = FALSE,
  verbose = TRUE,
  bug_name = NULL,
  plot_ext = "pdf"
)
}
\arguments{
\item{tree_file}{either a path to a tree file readable by \code{\link[ape:read.tree]{ape::read.tree()}}
or an object of class "phylo" that is already read into R}

\item{meta_file}{either a data frame of metadata or a path to file containing
the metadata}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{out_dir}{if saving, directory where to save}

\item{omit_na}{logical indicating whether to omit incomplete cases}

\item{plot_ext}{extension to use when saving plots}
}
\description{
Plot a tree file, and show the outcome variable as a colored dot
  on the end of each tip.
}
\details{
Showing the covariates as color bar annotations isn't supported yet.
}
