% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_tree_with_post}
\alias{plot_tree_with_post}
\title{Plot a tree and the PGLMM posterior on phylogenetic effects}
\usage{
plot_tree_with_post(
  tree_file,
  meta_file,
  fit,
  covariates = c("age", "gender"),
  outcome = "crc",
  omit_na = FALSE,
  verbose = TRUE,
  labels,
  return_tree_df = FALSE
)
}
\arguments{
\item{tree_file}{either a path to a tree file readable by \code{\link[ape:read.tree]{ape::read.tree()}}
or an object of class "phylo" that is already read into R. Ignored if
\code{cor_mat} is supplied.}

\item{meta_file}{either a data frame of metadata or a path to file containing
the metadata}

\item{fit}{a pglmm fit from \code{anpan_pglmm()}}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{omit_na}{logical indicating whether to omit incomplete cases}

\item{labels}{the ordered tip labels from the tree}

\item{return_tree_df}{if true, return a list containing 1) the plot, 2) the
segment data frame, and 3) the labelled terminal segment data frame.
Otherwise, just return the plot.}
}
\value{
either the plot or (if return_tree_df = TRUE) a list containing the
  plot, the segment df, the terminal segment df, and the yrep df.
}
\description{
Plot a tree and the PGLMM posterior on phylogenetic effects
}
\details{
The whiskers of each box plot are the 90% posterior intervals, the
  box is the 50% interval, and the middle line is the posterior mean.
}
