% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{read_and_filter}
\alias{read_and_filter}
\title{Read and filter a gene family file}
\usage{
read_and_filter(
  bug_file,
  metadata,
  pivot_wide = TRUE,
  minmax_thresh = 5,
  covariates = NULL,
  outcome = NULL,
  genomes_file = NULL,
  filtering_method = "kmeans",
  discretize_inputs = TRUE,
  discard_poorly_covered_samples = TRUE,
  save_filter_stats = TRUE,
  filter_stats_dir = NULL,
  plot_ext = "pdf",
  verbose = TRUE
)
}
\arguments{
\item{bug_file}{path to a gene family file (usually probably from HUMAnN)}

\item{pivot_wide}{logical indicating whether to return data in wide format}

\item{minmax_thresh}{genes must have at least this many (or N - this many)
non-zero observations or else be discarded}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{genomes_file}{optional file giving gene presence/absence of representative isolate genomes}

\item{filtering_method}{either "kmeans" or "none"}

\item{discretize_inputs}{logical indicating whether to discretize the input abundance
measurements (0/nonzero --> FALSE/TRUE) before passing them to the modelling function}

\item{discard_poorly_covered_samples}{logical indicating whether to discard samples where the genes of a bug are poorly covered}

\item{save_filter_stats}{logical indicating whether to save filter statistics}

\item{filter_stats_dir}{directory to save filtering statistics to}

\item{plot_ext}{extension to use for plots}
}
\description{
Read and filter a gene family file
}
