% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_data.R
\name{read_bug}
\alias{read_bug}
\title{Read a bug's genefamily file}
\usage{
read_bug(bug_file, meta = NULL, remove_pattern = "_Abundance-RPKs")
}
\arguments{
\item{bug_file}{path to a bug's genefamily file}

\item{meta}{a data frame of metadata}

\item{remove_pattern}{pattern to remove from the column names of the
genefamily file}
}
\value{
The genefamily file of the bug as a data.table in TALL format
}
\description{
Read a bug's genefamily file
}
\details{
The input bug_file needs to be readable by data.table::fread()

If metadata is provided, the genefamily file is subset to only those samples
present in the sample_id column of the metadata.
}
