% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{scone}
\alias{scone}
\title{Run scone}
\usage{
scone(
  bug_file,
  meta_file,
  out_dir,
  model_type = "scone",
  covariates = c("age", "sex")
)
}
\arguments{
\item{bug_file}{path to a gene family file (usually probably from HUMAnN)}

\item{meta_file}{path to a metadata tsv. Must contain the specify covariates}

\item{model_type}{either "scone" or "glm"}

\item{covariates}{covariates to account for}
}
\description{
Run the scone model on a
}
