% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{anpan}
\alias{anpan}
\title{Run anpan}
\usage{
anpan(
  bug_file,
  meta_file,
  out_dir,
  model_type = "anpan",
  covariates = c("age", "gender"),
  filtering_method = "kmeans",
  annotation_file = NULL,
  plot_ext = "png",
  save_filter_stats = TRUE
)
}
\arguments{
\item{bug_file}{path to a gene family file (usually probably from HUMAnN)}

\item{meta_file}{path to a metadata tsv. Must contain the specify covariates}

\item{model_type}{either "anpan" or "glm"}

\item{covariates}{covariates to account for CURRENTLY IGNORED}

\item{save_filter_stats}{logical indicating whether to save filter statistics}
}
\description{
Run the anpan model on a single bug
}
