% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{anpan}
\alias{anpan}
\title{Run anpan}
\usage{
anpan(
  bug_file,
  meta_file,
  out_dir,
  model_type = "anpan",
  covariates = c("age", "gender"),
  outcome = "crc",
  filtering_method = "kmeans",
  annotation_file = NULL,
  plot_ext = "png",
  save_filter_stats = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{bug_file}{path to a gene family file (usually probably from HUMAnN)}

\item{meta_file}{path to a metadata tsv}

\item{model_type}{either "anpan" or "glm"}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{save_filter_stats}{logical indicating whether to save filter statistics}
}
\description{
Run the anpan model on a single bug
}
\details{
The specified metadata file must contain columns matching "sample_id"
  and the specified covariates and outcome variables.
}
