% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{anpan_batch}
\alias{anpan_batch}
\title{Apply anpan to a many bugs}
\usage{
anpan_batch(
  bug_dir,
  meta_file,
  out_dir,
  model_type = "horseshoe",
  covariates = c("age", "gender"),
  outcome = "crc",
  omit_na = FALSE,
  filtering_method = "kmeans",
  discard_absent_samples = TRUE,
  annotation_file = NULL,
  save_filter_stats = TRUE,
  verbose = TRUE,
  plot_results = TRUE,
  plot_ext = "png",
  q_threshold = NULL,
  n_top = 50,
  ...
)
}
\arguments{
\item{bug_dir}{a directory of gene family files}

\item{meta_file}{path to a metadata tsv}

\item{out_dir}{path to the desired output directory}

\item{model_type}{either "horseshoe", "glm", or "fastglm"}

\item{covariates}{character vector of covariates to include in the model}

\item{outcome}{character string of the outcome variable}

\item{omit_na}{logical indicating whether to omit incomplete cases}

\item{discard_absent_samples}{logical indicating whether to discard samples when a bug is labelled as completely absent}

\item{annotation_file}{optional path file giving annotations}

\item{save_filter_stats}{logical indicating whether to save filter statistics}

\item{plot_results}{logical indicating whether or not to plot the results}

\item{plot_ext}{extension to use for plots}

\item{q_threshold}{FDR threshold to use for inclusion in the plot.}

\item{n_top}{number of top elements to show from the results}

\item{...}{arguments to pass to brms::brm()}
}
\description{
This function calls anpan() on each gene family file in the
  \code{bug_dir} directory and makes a composite data + results plot for
  each.
}
