% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.R
\name{anpan_pglmm}
\alias{anpan_pglmm}
\title{Run a phylogenetic generalized linear mixed model}
\usage{
anpan_pglmm(
  meta_file,
  tree_file,
  outcome,
  out_dir = NULL,
  trim_pattern = NULL,
  covariates = NULL,
  omit_na = FALSE,
  family = "gaussian",
  plot_cor_mat = TRUE,
  save_object = FALSE,
  verbose = TRUE,
  test_signal = TRUE,
  reg_noise = TRUE,
  ...
)
}
\arguments{
\item{meta_file}{path to a metadata tsv}

\item{tree_file}{path to a .tre file}

\item{outcome}{the name of the outcome variable}

\item{out_dir}{if saving, directory where to save}

\item{trim_pattern}{optional pattern to trim from tip labels of the tree}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{omit_na}{logical indicating whether to omit incomplete cases}

\item{family}{string giving the name of the distribution of the outcome
variable (usually "gaussian" or "binomial")}

\item{save_object}{logical indicating whether to save the model fit object}

\item{test_signal}{compute the phylogenetic signal}

\item{reg_noise}{logical indicating whether to regularize the ratio of
sigma_phylo to sigma_resid with a Gamma(1.33,2) prior}

\item{...}{other arguments to pass to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}}
}
\description{
Run a phylogenetic generalized linear mixed model
}
\details{
the tip labels of the tree must be the sample ids from the metadata.
You can use the \code{trim_pattern} argument to automatically trim off any
consistent pattern from the tip labels if necessary.

The dots can be used to pass e.g. parallel_chains=4 to make the chains run
in parallel.

The prior for the intercept is a normal distribution centered on the mean
of the outcome variable with a standard deviation of 3*sd(outcome variable)

The default error distribution for the outcome is "gaussian". You could
change this to a phylogenetic logistic regression by changing \code{family}
to "binomial" for example.
}
\seealso{
\code{\link[loo:loo]{loo::loo()}}, \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}
}
