% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.R
\name{anpan_pglmm}
\alias{anpan_pglmm}
\title{Run a phylogenetic generalized linear mixed model}
\usage{
anpan_pglmm(
  meta_file,
  tree_file = NULL,
  cor_mat = NULL,
  outcome,
  covariates = NULL,
  out_dir = NULL,
  trim_pattern = NULL,
  bug_name = NULL,
  omit_na = FALSE,
  ladderize = TRUE,
  family = "gaussian",
  show_plot_cor_mat = TRUE,
  show_plot_tree = TRUE,
  show_post = TRUE,
  show_yrep = FALSE,
  save_object = FALSE,
  verbose = TRUE,
  loo_comparison = TRUE,
  run_diagnostics = TRUE,
  reg_noise = TRUE,
  reg_gamma_params = c(1, 2),
  plot_ext = "pdf",
  beta_sd = NULL,
  int_prior_scale = 1,
  sigma_phylo_scale = 0.333,
  ...
)
}
\arguments{
\item{meta_file}{either a data frame of metadata or a path to file containing the metadata}

\item{tree_file}{either a path to a tree file readable by \code{\link[ape:read.tree]{ape::read.tree()}} or an object of
class "phylo" that is already read into R. Ignored if \code{cor_mat} is supplied.}

\item{cor_mat}{a correlation matrix provided as an alternative to a tree.}

\item{outcome}{the name of the outcome variable}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{out_dir}{if saving, directory where to save}

\item{trim_pattern}{optional pattern to trim from tip labels of the tree}

\item{omit_na}{logical indicating whether to omit incomplete cases of the metadata}

\item{ladderize}{logical indicating whether to run \code{\link[ape:ladderize]{ape::ladderize()}} on the tree before running
the model}

\item{family}{string giving the name of the distribution of the outcome variable (usually
"gaussian" or "binomial")}

\item{show_plot_cor_mat}{show a plot of the correlation matrix derived from the tree}

\item{show_plot_tree}{show a plot of the tree overlaid with the outcome.}

\item{show_post}{show a plot of the tree overlaid with the outcome and posterior distribution on
phylogenetic effects.}

\item{show_yrep}{show a plot of the tree overlaid with the outcome and the posterior predictive
distribution for each observation if plotting the tree}

\item{save_object}{logical indicating whether to save the model fit object}

\item{loo_comparison}{logical indicating whether to compare the phylogenetic model against a base
model (without the phylogenetic term) using \code{\link[loo:loo_compare]{loo::loo_compare()}}}

\item{run_diagnostics}{logical indicating whether to run \code{\link[cmdstanr:fit-method-cmdstan_summary]{cmdstanr::cmdstan_diagnose()}} and
\code{\link[loo:pareto-k-diagnostic]{loo::pareto_k_table()}} to check the MCMC and loo diagnostics respectively.}

\item{reg_noise}{logical indicating whether to regularize the ratio of sigma_phylo to sigma_resid
with a Gamma prior}

\item{reg_gamma_params}{the shape and rate parameters of the Gamma prior on the noise term ratio.
Default: c(1,2)}

\item{plot_ext}{extension to use when saving plots}

\item{beta_sd}{prior standard deviation parameters on the normal distribution for each covariate
in the GLM component}

\item{int_prior_scale}{standard deviation of the 0-centered normal prior for the intercept of the
model (with centered covariates)}

\item{sigma_phylo_scale}{standard deviation of half-normal prior on \code{sigma_phylo} for
logistic PGLMMs when \code{family = 'binomial'}. Increasing this value can easily lead to
overfitting.}

\item{...}{other arguments to pass to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}}
}
\value{
A list containing the model input (in the order passed to the model), estimated
correlation matrix, the pglmm fit object, and (if \code{loo_comparison} is on) the base fit
object and the associated loo objects.
}
\description{
Run a phylogenetic generalized linear mixed model
}
\details{
the tip labels of the tree must be the sample ids from the metadata. You can use the
\code{trim_pattern} argument to automatically trim off any consistent pattern from the tip
labels if necessary.

The default error distribution for the outcome is "gaussian". You could change this to a
phylogenetic logistic regression by changing \code{family} to "binomial" for example.

The prior for the intercept is a normal distribution centered on the mean of the outcome
variable with a standard deviation of 3*sd(outcome variable).

If a prior scale on covariate effects isn't specified (i.e. \code{beta_sd} is left at the
default NULL), the prior scale is set to 1 / (1 standard deviation) for each covariate. The
values will be shown in a message. It's preferable to set \code{beta_sd} with domain knowledge.
Consider the units/scale of your outcome variable if using \code{family = "gaussian"} or the
range of plausible coefficient values on the log-odds scale if using \code{family =
  "binomial"}.

If supplying \code{beta_sd} with a categorical predictor with >2 levels, only specify a single
corresponding element in beta_sd. This appropriate element will get repeated as necessary.

\code{sigma_phylo_scale} is used to set the width of the half-normal prior on sigma_phylo when
\code{family = "binomial"} (no effect when \code{family = "gaussian"}). Increasing this value
can easily lead to overfitting. If sigma_phylo is high, then leafwise phylogenetic effects can
get very far from zero (disregarding the correlation structure and covariate effects), which
lets the model perfectly fit each observation. The default value keeps most of the prior mass
on leafwise phylogenetic effects between +/- 1, which is still pretty liberal on the log-odds
scale.

The model indexes the leaves according to the Cholesky factor of the correlation matrix. This
likely won't be the same order as the input tree. The \code{model_input} that is returned as
part of the output will have the samples in a factor with the model index order if needed.

It's normal to see some warnings during warmup, particularly about "Scale vector is inf".

This function tries to get the \code{bug_name} argument from the tree file, but if it's not
provided you may need to set it yourself.

Common cmdstanr options that one might want to pass in via ... include \code{refresh = 500}
(show fewer MCMC progress updates), \code{adapt_delta = .98} (avoid divergences at the cost of
possibly needing more iterations to get convergence), \code{show_messages = FALSE} ,  and \code{parallel_chains = 4} (run the
MCMC chains in parallel).

If you want to use the PGLMM log-likelihood data frame with functions from the \code{loo}
package, you'll need to convert it to a matrix with \code{as.matrix()}. It's converted to a
tibble internally so that it prints nicely.

If \code{int_prior_scale} isn't specified, it defaults to 1 for binary outcomes and 1 standard
deviation of the outcome for gaussian outcomes.
}
\examples{
meta = data.frame(x = rnorm(100), sample_id = paste0("t", 1:100))
tr = ape::rtree(100)
anpan_pglmm(meta, tr,
outcome = "x",
iter_sampling = 10, # Use more in practice!
iter_warmup = 10,
show_plot_cor_mat = FALSE,
show_plot_tree = FALSE)
}
\seealso{
\code{\link[=anpan_pglmm_batch]{anpan_pglmm_batch()}}, \code{\link[loo:loo]{loo::loo()}}, \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}
}
