% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglmm.R
\name{anpan_pglmm}
\alias{anpan_pglmm}
\title{Run a phylogenetic generalized linear mixed model}
\usage{
anpan_pglmm(
  meta_file,
  tree_file,
  outcome,
  out_dir = NULL,
  trim_pattern = NULL,
  covariates = NULL,
  bug_name = NULL,
  omit_na = FALSE,
  family = "gaussian",
  show_plot_cor_mat = TRUE,
  show_plot_tree = TRUE,
  save_object = FALSE,
  verbose = TRUE,
  loo_comparison = TRUE,
  reg_noise = TRUE,
  plot_ext = "pdf",
  show_yrep = TRUE,
  ...
)
}
\arguments{
\item{meta_file}{either a data frame of metadata or a path to file containing
the metadata}

\item{tree_file}{either a path to a tree file readable by \code{\link[ape:read.tree]{ape::read.tree()}}
or an object of class "phylo" that is already read into R}

\item{outcome}{the name of the outcome variable}

\item{out_dir}{if saving, directory where to save}

\item{trim_pattern}{optional pattern to trim from tip labels of the tree}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{omit_na}{logical indicating whether to omit incomplete cases}

\item{family}{string giving the name of the distribution of the outcome
variable (usually "gaussian" or "binomial")}

\item{show_plot_cor_mat}{show a plot of the correlation matrix derived from
the tree}

\item{show_plot_tree}{show a plot of the tree overlaid with the outcome}

\item{save_object}{logical indicating whether to save the model fit object}

\item{loo_comparison}{logical indicating whether to compare the phylogenetic
model against a base model (without the phylogenetic term) using
\code{\link[loo:loo_compare]{loo::loo_compare()}}}

\item{reg_noise}{logical indicating whether to regularize the ratio of
sigma_phylo to sigma_resid with a Gamma(1,2) prior}

\item{plot_ext}{extension to use when saving plots}

\item{show_yrep}{logical indicating whether to show the posterior predictive
distribution for each observation if plotting the tree}

\item{...}{other arguments to pass to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}}
}
\description{
Run a phylogenetic generalized linear mixed model
}
\details{
the tip labels of the tree must be the sample ids from the metadata.
You can use the \code{trim_pattern} argument to automatically trim off any
consistent pattern from the tip labels if necessary.

The dots can be used to pass e.g. parallel_chains=4 to make the chains run
in parallel.

The prior for the intercept is a normal distribution centered on the mean
of the outcome variable with a standard deviation of 3*sd(outcome variable)

The default error distribution for the outcome is "gaussian". You could
change this to a phylogenetic logistic regression by changing \code{family}
to "binomial" for example.

It's normal to see some warnings during warmup, particularly about "Scale
vector is inf".

This function tries to get the \code{bug_name} argument from the tree file,
but if it's not provided you may need to set it yourself.
}
\seealso{
\code{\link[loo:loo]{loo::loo()}}, \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}
}
