% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwy_ranef.R
\name{anpan_pwy_ranef_batch}
\alias{anpan_pwy_ranef_batch}
\title{Fit the pathway random effects model for multiple bugs}
\usage{
anpan_pwy_ranef_batch(
  bug_pwy_dat,
  group_ind = "crc",
  out_dir = NULL,
  group_exp_rate = 3,
  ...
)
}
\arguments{
\item{bug_pwy_dat}{a data frame with a row for each observation and columns "pwy",
"log10_species_abd", "log10_pwy_abd", and a group indicator column named according to the
\code{group_ind} argument}

\item{group_ind}{a character giving the name of the column for the 0/1 group indicator variable
in \code{bug_pwy_dat}}

\item{out_dir}{output directory}

\item{group_exp_rate}{rate parameter of the exponential distribution prior on group effects}

\item{...}{other arguments to pass to cmdstanr::sample()}
}
\value{
a tibble of row-binded anpan_pwy_ranef results
}
\description{
Fit the pathway random effects model for multiple bugs
}
\details{
In addition to the column requirements of \code{anpan_pwy_ranef()}, the input data frame
  \code{bug_pwy_dat} here must also contain a variable called "bug" which gives a unique
  identifier for each bug.
}
\examples{
\dontrun{
library(tidyverse)
library(anpan)

set.seed(123)
input_dat = tibble(bug = rep(paste0("bug", 1:5), each = 200),
                   pwy = rep(paste0('pwy', 1:5), times = 200),
                   log10_species_abd = rnorm(1000),
                   log10_pwy_abd = rnorm(1000, mean = .8*log10_species_abd),
                   group = sample(c(0,1), size = 1000, replace = TRUE))
 # ^ the pathway and and group are NOT related in any pathway.

res = anpan_pwy_ranef_batch(input_dat, group_ind = "group")

# Examine the summary
pwy_group_res = res |>
  select(bug, summary_df) |>             # select the two main columns
  unnest(c(summary_df)) |>               # unnest
  filter(grepl("^pwy_eff", variable)) |> # get just the pwy:group terms
  arrange(-abs(mean))                    # sort by decreasing effect size

print(pwy_group_res)

pwy_group_res |> filter(hit)
# ^ Here, there are no hits because we simulated with no dependence
}

}
\seealso{
\code{\link[=plot_pwy_ranef]{plot_pwy_ranef()}} , \code{\link[=plot_pwy_ranef_intervals]{plot_pwy_ranef_intervals()}}, \code{\link[cmdstanr:CmdStanMCMC]{cmdstanr::CmdStanMCMC}}
}
