% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_results}
\alias{plot_results}
\title{Plot the data for top results}
\usage{
plot_results(
  res,
  covariates,
  outcome,
  model_input,
  discretize_inputs = TRUE,
  plot_dir = NULL,
  bug_name,
  annotation_file = NULL,
  cluster = "none",
  show_trees = FALSE,
  n_top = 50,
  q_threshold = NULL,
  show_intervals = TRUE,
  width = NULL,
  height = NULL,
  plot_ext = "pdf"
)
}
\arguments{
\item{res}{a data frame of model results (from \code{anpan}) for the genes of a single bug (i.e.
the output written to *gene_terms.tsv.gz)}

\item{covariates}{character string of the covariates to show}

\item{outcome}{character string of the outcome variable}

\item{model_input}{data frame of the model input}

\item{plot_dir}{directory to write output to}

\item{bug_name}{character string giving the name to use in the title/output file}

\item{annotation_file}{optional path file giving annotations}

\item{cluster}{axis to cluster. either "none", "samples", "genes", or "both"}

\item{show_trees}{logical to show the trees for the samples (if clustered)}

\item{n_top}{number of top elements to show from the results}

\item{q_threshold}{FDR threshold to use for inclusion in the plot.}

\item{width}{width of saved plot in inches}

\item{height}{height of saved plot in inches}

\item{plot_ext}{extension to use for plots}
}
\description{
This funciton makes a tile plot of the top results of a fit alongside another tile
  plot showing the covariates included. Optional annotations can be included.
}
\details{
If included, \code{annotation_file} must be a tsv with two columns: "gene" and
  "annotation".

  \code{n_top} is ignored if \code{q_threshold} is specified.

  When \code{cluster = "none"}, the samples are ordered by metadata and the genes are ordered by
  statistical significance.

  When signficance stars are shown, they encode the following (fairly standard) signficance
  thresholds: p.value < .001 ~ ***, p.value < .01  ~ **, p.value < .05  ~ *, p.value < .1   ~ .,
  p.value < 1    ~ " "

  If applicable, the Q-value used to color the dot on the interval panel is q_global if present
  in the input and q_bug_wise otherwise.
}
