% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_tree_with_post_pred}
\alias{plot_tree_with_post_pred}
\title{Plot a tree and the PGLMM posterior predictive}
\usage{
plot_tree_with_post_pred(
  tree_file,
  meta_file,
  covariates = c("age", "gender"),
  outcome = "crc",
  omit_na = FALSE,
  verbose = TRUE,
  fit = NULL,
  labels
)
}
\description{
Plot a tree and the PGLMM posterior predictive
}
