% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{read_and_filter}
\alias{read_and_filter}
\title{Read and filter a gene family file}
\usage{
read_and_filter(
  bug_file,
  metadata,
  pivot_wide = TRUE,
  minmax_thresh = 5,
  covariates = NULL,
  outcome = NULL,
  filtering_method = "kmeans",
  discard_absent_samples = TRUE,
  save_filter_stats = TRUE,
  filter_stats_dir = NULL,
  plot_ext = "pdf",
  verbose = TRUE
)
}
\arguments{
\item{bug_file}{path to a gene family file (usually probably from HUMAnN)}

\item{minmax_thresh}{genes must have at least this many (or N - this many)
non-zero observations or else be discarded}

\item{covariates}{covariates to account for (as a vector of strings)}

\item{outcome}{the name of the outcome variable}

\item{discard_absent_samples}{Discard samples that are labelled as missing
the bug according to the filtering strategy (otherwise keep them as
observations of absence)}

\item{save_filter_stats}{logical indicating whether to save filter statistics}
}
\description{
Read and filter a gene family file
}
